Add cards (#4)
* Add all images for base and "child" cards

"Child" cards refer to those that are added by the "add to your hand" ability.

We will add the "child" cards from the "spawn" ability later

* Replace "Obtain" column with "Image" column

Image column refers to the filename of the image

* Add images for summon cards

* Add a few card images I missed out on + SQL query

Python script for running SQL query to insert all these extra cards

* Added Child Cards

Created "Extra Cards" table and added child cards to it.

Changed "Moogle Trio" to "Moogle Bard"

* Removed outdated DB

* Use SQL query to update name of card with ID 149

Was called "Moogle Trio" instead of "Moogle Bard"

* Add script for reading card ranges

For each card:
1. Crops the 5x5 grid from the card image
2. Get the colour of each square.
3. Map colour to a 5x5 array

* Fix name of levrikon image file and tweak script

- Account for "Mk. II" -> "mk-ii"
- Account for different naming format for "spawned" cards
- Tweak border pixel

* Rename scraper folder to scripts

* Add script for creating `Range` table

This table maps the "significant" tiles which are coloured orange, red, or both orange and red to the card's ID.

Additionally, we also include the offset relative to the position the card is for easier calculations.

---------

Co-authored-by: Kai <140022059+frankpadada@users.noreply.github.com>