Add abilities and tests (#9)
* Add redirect when trying to change game URL

If the user is not in a game and they try modifying their URL to e.g "/game/DXyszP" they will be redirected back to the home page

* Create back button in deck page

Also use the default deck if there is no saved deck in localStorage

Also created xUnit test project

* Add unit tests for game start features

- Tests are done with each player having the default deck
- Each players must have 15 starting cards
- When game starts, draw 5 to hand

* Add board initialization and unit tests for it

* Add unit test for placing Security Officer

* Add gitignore for xUnit project

* Update card initialization

* Refactor unit test for checking tile state

- Make it more compact
- Also added a few event listeners. They don't do anything yet.

* Add unit test for card with +R ability

Tested with the "Twin Brain" card

* Add ability for cards with AP and EP condition

The implementation's a bit messy right now but appears to be functional

* Add enhance cards on affected tiles ability

 - Testing Crystalline Crab's ability (*) (+P) and removing its effects when removed
- Testing Loveless' ability (P) (+P)

* Add ability to self-enhance on enhance/enfeeble

This includes cards with the Target of "s" and a condition of +A,-A,+E,-E,-AE, or +AE

* Add test for cards with AE condition

- This includes fixes to make it work correctly
- Also a test for card that executes ability when it is destroyed (post mortem)

* Fixes and tests for when first enhanced abilities

* Introduce CardBonusPower and TileBonusPower

There are now 3 distinct forms of power bonuses.

- TileBonusPower refers to the bonus power due to cards with "While in play" condition
- CardBonusPower refers to the bonus power due to other abilities affecting this card which don't have the the "While in play" condition
- SelfBonusPower refers to the bonus power for cards with ability to enhance their own power

* Add ability with AED trigger condition

A few tweaks to get it to work

* Update script for making parent-child table

The `Extra Card` table we had before didn't have a proper link between the parent and child.

The FK was just NULL

I also fixed the Data Source for the database in the unit tests

* Add unit tests for cards with "add" ability

* Implement "spawn" ability

As part of the implementation, I had to decouple the SwapPlayerTurns() from the PlaceCard() method

* Refactor code and add win lane bonus ability

* Separate tests into individual files & L+V

-L+V is the ability for the "Ultimate Party Animal" card.
- Also slightly refactored the "+Score" ability implementation

* Add destroy card ability

Added a unit test using Sephiroth card

* Add Cloud's ability

Only card with the P1R trigger condition. Added unit test too

* Add the ability for "Two Face" card

- Different actions based on whether it was enhanced or enfeebled
- Unit test for both
- New game logic for queuing actions (e.g. wait for a card to finish processing its tiles and abilities before invoking the abilities of other cards affected by it)

* Add test for picking up card