Preview Move Backend Logic & Frontend Mulligan Screen (#11)
* Add feature to deep copy a game instance

- This is used for generating a copy to perform an action to preview
- In order to deep copy, I had to SERIALIZE --> DESERIALIZE the original game object

* Refactor to send DTO instead of whole Card

Also refactored to have less repeated code with populating the cards in GameRepository.cs and TestBase.cs

* Improve design of mulligan screen

- Did a lot of refactoring

* Refactor CardInHand and MulliganCard

Created a common Card component to reduce duplicate code

Problem now is that I have 3 files that export a "Card"

* Split functionality of <Board /> component

Before it conditionally rendered the mulligan phase or or the actual game. Splitting it up makes it easier to read

* Outline cards of the player whose turn it is

The `ReadyUp()` signalR method now sends the id of the player who is currently playing.

My SignalRProvider has a `playing` state which is the id of the player who has the turn

* Add drag and drop function

Board isn't properly implemented. Just a simple drag-and-drop to a box and no further functionality