Improve "hitbox" of the cards in hand
I was using display: flex before to try align the cards which were positioned absolutely.

Consequently, there was a mismatch between what we see and where the draggable component actually is.

This time around, I don't rely on a flexbox and just position the components absolute to the hand component. This resolves the issue before.