Fix target abilities affecting wrong player's card
I had a logic error:
var player = tile.Owner ?? instigator!;
Does not work in all cases. Instead I had to write a switch statement to make the decision on which players are affected based on the ability.target attribute.

There is currently a MAJOR bug with the preview. Deep copying is fundamentally flawed because we lose our references for the Actions. I might need to seek an alternative implementation