Change UPA's ability to trigger at each round
Before it was triggering only at the end of the game. Now it shows the loser's points as bonus power for the victor.

I also fixed a bug where self-enhancing cards enhance themselves when they card has been enhanced. This is due to CalculateSelfBoostFromPowerModifiedCards() being called when a card is played if it is a self-enhancing card.