Fix adding extra card to EnhancedCard in preview
I used .Contains() which takes in a Tile object as an argument. When cloning the tiles in EnhancedCards no longer reference the tile in the grid.

I fixed this by assigning each tile an ID which corresponds to the location it is in Player1Grid. With this, I can remap the tiles in EnhancedCards/EnfeebledCards to the tile in the grid.