Fix self enhance ability and replace card (#13)
* Add feature to track changes in EnhancedCards

* Fix self-enhancing cards recursion/outdated

Add them to EnhanceCards list but leave the calculations till the end so that when enhancing, it has the most up to date information

* Fix destroying amalgam giving two cards`

Result of having two cards subscribed to OnCardDestroyed.

Resolved by making sure it is the card being destroyed before executing the "post-mortem" ability.

* Fix: placing card w/ power < enfeeble not destroy

Cards with a cumulative power of 0 or less should be destroyed

* Fix adding extra card to EnhancedCard in preview

I used .Contains() which takes in a Tile object as an argument. When cloning the tiles in EnhancedCards no longer reference the tile in the grid.

I fixed this by assigning each tile an ID which corresponds to the location it is in Player1Grid. With this, I can remap the tiles in EnhancedCards/EnfeebledCards to the tile in the grid.

* Fix self-enhancing when empty tile is enhanced

An empty tile getting enhanced/enfeebled should not invoke OnEnfeebledCardsChanged/OnEnhancedCardsChanged.

Also fixed RemoveFromEnfeebledCards() and AddToEnfeebledCards(). Before I wasn't using the ActionQueue and was invoking OnEnhancedCardsChanged

* Fix not subbed to OnEnfeebledCardsChanged

Accidentally subscribed to OnEnhancedCardsChanged

* Fix correct EnhancedCards, wrong self-enhance

A special case when an enemy plays a card that enhances while in play, when an allied card is on the affected tile.

The tile should go into EnhancedCards, but when calculating self-enhance bonus power, don't include it as the bonus power is ONLY for the enemy player.

* Fix preview move breaking when card enfeebled

* Fix uninit ability using incorrect grid

Using a different grid will affect which way the card's range is pointed to. During un-initialization, it will lead to the card making changes to tiles which shouldn't be affected, instead of the ones which should change.

* Fix accidentally removing card from EnhancedCards

I had a seemingly useless else statement which removes a card from both EnhancedCards and EnfeebledCards

* Fix Sandspitter ability

Data scraped from game8 for Sandspitter is incorrect. I manually changed it using data from the IGN website

* Fix destroying source not updating self-enhance

When a self-enhance card no longer has any of its dependencies, it should reset its SelfBonusPower to 0

* Fix self-enhance card placed on enhanced tile

There was a logic error where it would remove itself from the EnhancedCards list if it was the only enhanced card. It could be enhanced by another card, but would still remove itself from the list.

* Fix enhance on card destroyed preview not working

I was comparing by reference instead of Owner.Id

* Refactor & fix issues with replacing cards

Replacing a card which is enhanced due to a tile bonus removes it from EnhancedCards when it shouldn't

Replacing a card and enfeebling affected card by replaced card's power had some weird side effects